Nbt -> NBT
